Add files via upload
Task: design tests in the style of unittest 
Try to style the tests from the first module in unittest style.

Take tests from step - https://stepik.org/lesson/138920/step/11?unit=196194
Create a new file
Create a class with tests in it, which should be inherited from unittest.TestCase by analogy with the previous step
Rewrite unittest style test for page  http://suninjuly.github.io/registration1.html
Rewrite in unittest style the second test for the page  http://suninjuly.github.io/registration2.html
Decorate final checks in unittest style tests, for example using the assertEqual check method
Run resulting tests from file 
Look at the launch report and find the last line 
Submit this line as an answer to this assignment 
Note that the intention is to throw a NoSuchElementException in the second test. If you used the try/except construct, you need to run the test without it here. Catching exceptions is not necessary (at least here)!

This is all to illustrate that unittest will run the tests and summarize the results even in the presence of an unexpected exception. 

Do not delete the code after completing this task, it will come in handy in the next lesson.